ELF unexec: _OBJC_ symbols in bss sections
This code assumed that there was only one bss section.  Rather than
checking for a particular index, check the section type.  Also, handle
the possibility that the section was SHT_NOBITS originally and is
unchanged, in which case no clearing is needed (and sh_offset isn't
necessarily valid, which can lead to a wild memset).

* unexelf.c (unexec): Properly handle _OBJC_ symbols in bss sections.