Remove packages obsoleted before Emacs 24.
In accordance with the policy discussed in the emacs-devel list,
packages that have been obsoleted for a full major release cycle are up
for deletion.

This removes almost all packages that are now eligible for deletion,
with the exception of "cl-compat", which seems it is likely to still be
used, and "optional", which offers some functionality that doesn't have
a replacement yet.