comint-insert-previous-argument doesn't detect and ignore trailing &
This function is invoked in shell-mode by the user, and is meant to
emulate what M-. does in zsh and bash: it inserts an argument from a
previous command.  Neither zsh nor bash treat a trailing & specially:
M-. simply inserts it if it is encountered.  Emacs DID have extra
logic to detect and discard trailing &, but this logic was buggy, and
a && anywhere in the sequence would confuse it.  This patch simply
removes that logic to fix the bug and to emulate zsh and bash more
closely

* lisp/comint.el (comint-insert-previous-argument): don't detect and
  ignore trailing &
  (Bug#25271)
* etc/NEWS: Document this.