Make tasks markdown checkable.
Also added a bit of instructions on creating a new issue when someone picks a task. The idea is we can use README to keep low hanging fruit available for people. Another option (maybe later, but much less visible)  would be to create issues for each task and tag them as unasigned, by complexity and such. 

By now I guess it's a good idea to have them (while they are few) discoverable on the readme.