Integrate cargo with automake
Adds appropriate rules and dependencies to the existing autoconf/make
based build system. This includes the build process itself, make
clean and make check, obviating the need to manually edit the
makefiles to account for differnt build configurations.

The default optimisation level is --release in line with the C
components however a configure option (--enable-rust-debug) to allow
this to be overriden in isolation.

Arguments may be passed to cargo via the CARGO_FLAGS variable.

The RUSTFLAGS environment variable is also explicitly passed when
cargo build is invoked.