Updating implementation for rust_make_temp.
This implementation is removing the dep. of "tempfile", as it does
not suit our needs for remacs. Instead we 'roll or own' for the small
amount of functionality we need.

As of this commit, we will need an implementation of the windows
file creation, and of an error code implementation.