Fix docstring of dabbrev-abbrev-char-regexp
* lisp/dabbrev.el (dabbrev-abbrev-char-regexp): Using a value of nil
is equivalent to "\\sw\\|\\s_", and has no special behavior.  If the
previous character doesn't match, we search backwards for one that
does, not throw an error.  Replace Lisp example with C based one to
make it clear that "symbol" means a sequence of word and symbol
constituent characters, not a Lisp symbol (Bug#358).