Speed up the visual-mode relative line numbers
* src/xdisp.c (display_count_lines_visually): Introduce a
shortcut: if a relative line number was already calculated for
this iterator object, just increase it instead of the
expensive call to move_it_to.  Argument list changed to pass a
pointer to the iterator object.
(maybe_produce_line_number): Adjust for change in signature of
display_count_lines_visually.  Record the relative line number and
the corresponding byte position in the iterator object also in the
'visual' mode.