Use Travis build stages
This should help new contributors, who are often caught out by Travis
requiring rustfmt. Treat rustfmt as a separate stage, so it's clear
when the build failure is just a style issue.

Whilst build stages are a beta feature:
https://docs.travis-ci.com/user/build-stages/ they have been available
for several months, since May 2017:
https://blog.travis-ci.com/2017-05-11-introducing-build-stages

We also move configure into the before_script stage, so Travis shows
it separately in the logs. This should make the logs more organised
and readable.