Expand recognized time intervals for MPC seeking
Now accepts [+-]H:M:S and subsets.  Also accepts some odd variations of
it since the regexp is not strict.  One unpleasant caveat is that
string-to-number simply returns zero on failure instead of signaling an
error.  At the moment, there are cases where instead of getting
a user-error, the seek may simply not go where one expects it.
* lisp/mpc.el (mpc-read-seek): New function.
(mpc-seek-current): Use it.