make the REmacs icon distinct from the default Emacs icon
Please excuse the crudity of this model. I didn't have time to build
it to scale or paint it...

Err, actually painting it, is the only thing I did.  Also it took
Emacs a long time to move over from using ugly design, and I am afraid
this might be considered a regression.

If there is someone knowing what he/she is doing, please feel free to
make this nicer.

Also I will not be insulted if this PR would be rejected.