DRY is always better. (#456)
Computers are for automation. Let's apply their boredom.

Use a macro to wrap `defsubr`

The goal is to allow changes to calls of `defsubr` to be made in one
place.

Generate the `rust_init_syms call based on whether
a file has the `export_lisp_fns` macro.

If a `no_mangle` function is defined then export it.

Replace explicit list in a macro with a generator for the macro call

The `export_lisp_fn` macro is still used. The developer does not have
to ensure the names are in the list. This can be determined by parsing
the code.