Save the server alias on reconnect (Bug#29657)
rcirc does not retain the server alias on reconnect.  As a result, rcirc
fails to re-use server and channel buffers when an alias is used.  Further
problems may ensue when aliases are used to differentiate multiple
connections to the same host, for example when using a single IRC bouncer
or proxy to connect to multiple IRC networks.

Save the server alias when connecting to a server so that reconnect will
retain the alias.
* lisp/net/rcirc.el (rcirc-connect): Include server-alias when setting
rcirc-connection-info.

Copyright-paperwork-exempt: yes