Automatically convert arguemts to native Rust types when possible
Not every type can currently be converted, but new types can be
handled simply by adding an impl From<LispObject> for them.

Return values are not currently handled, but could very easily be done
in the same way as arguments, by adding an impl From<foo> for
LispObject for all the relevant types. Some discussion of that will be
necessary, however.

The automatic conversions signal the usual "argument type error" if
the incoming arguments are of incompatible types. Functions which
aren't supposed to signal an error, or which need to handle multiple
argument types, should simply keep using LispObject.